split favorite from parking
