+++ lib/services/parking.ts
@@ -402,111 +402,3 @@ export async function rateParkingById(
     await prisma.$disconnect();
   }
 }
-
-export async function addToFavoriteParkingById(
-  parkingId: string,
-  userId: string
-): Promise<ServerFunctionResponse<TParking | null>> {
-  try {
-    const parking = (await getParkingById(parkingId)).data;
-    const user = (await getUserById(userId)).data;
-
-    if (!parking) {
-      return new ErrorServerFunctionResponse("Nie znaleziono parkingu");
-    }
-
-    if (!user) {
-      return new ErrorServerFunctionResponse("Nie znaleziono użytkownika");
-    }
-
-    const favoredParkingByUser = await prisma.favoriteParkingAndUser.findFirst({
-      where: {
-        AND: [
-          {
-            parkingId: parseInt(parkingId),
-          },
-          {
-            userId: parseInt(userId),
-          },
-        ],
-      },
-    });
-
-    if (favoredParkingByUser) {
-      return new ErrorServerFunctionResponse("Parking już jest w ulubionych");
-    }
-
-    await prisma.favoriteParkingAndUser.create({
-      data: {
-        parkingId: parking.id,
-        userId: user.id,
-      },
-    });
-
-    const favoriteParking = (await getParkingById(parkingId)).data;
-    if (!favoriteParking) {
-      return new ErrorServerFunctionResponse(
-        "Błąd dodawania parkingu do ulubionych"
-      );
-    }
-
-    return new SuccessServerFunctionResponse(
-      "Pomyślnie dodano parking do ulubionych",
-      favoriteParking
-    );
-  } catch (err: unknown) {
-    return new ExceptionServerFunctionResponse(getErrorMessage(err));
-  } finally {
-    await prisma.$disconnect();
-  }
-}
-
-export async function removeFromFavoriteParkingById(
-  parkingId: string,
-  userId: string
-): Promise<ServerFunctionResponse<null>> {
-  try {
-    const parking = (await getParkingById(parkingId)).data;
-    const user = (await getUserById(userId)).data;
-
-    if (!parking) {
-      return new ErrorServerFunctionResponse("Nie znaleziono parkingu");
-    }
-
-    if (!user) {
-      return new ErrorServerFunctionResponse("Nie znaleziono użytkownika");
-    }
-
-    const favoredParkingByUser = await prisma.favoriteParkingAndUser.findFirst({
-      where: {
-        AND: [
-          {
-            parkingId: parking.id,
-          },
-          {
-            userId: user.id,
-          },
-        ],
-      },
-    });
-
-    if (!favoredParkingByUser) {
-      return new ErrorServerFunctionResponse("Parking nie jest w ulubionych");
-    }
-
-    await prisma.favoriteParkingAndUser.delete({
-      where: {
-        id: favoredParkingByUser.id,
-      },
-    });
-
-    return new SuccessServerFunctionResponse(
-      "Pomyślnie usunięto parking z ulubionych",
-      null
-    );
-  } catch (err: unknown) {
-    return new ExceptionServerFunctionResponse(getErrorMessage(err));
-  } finally {
-    await prisma.$disconnect();
-  }
-}
